+++ applibrary/src/main/java/com/hoomanholding/applibrary/model/data/request/OrderToggleStateRequest.kt
@@ -10,5 +10,5 @@ import com.hoomanholding.applibrary.model.data.enums.EnumState
 data class OrderToggleStateRequest(
     val state: EnumState,
     val description: String?,
-    val orderIds: List<Long>
+    val orderIds: List<Int>
 )
